backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
psycopg[binary]==3.2.6
python-dotenv==1.0.1
pywebpush==2.0.3
websockets==14.2
